Working on more flexible structure for material information from collada - material resolve step used to happen during the collada compile step - now it is delayed to layer - collada compile exports unresolved ("raw") materials

+++ RenderCore/Assets/ModelRunTime.h
@@ -185,7 +185,7 @@ namespace RenderCore { namespace Assets
 
         static bool CanDoPrepareAnimation(Metal::DeviceContext* context);
 
-        std::vector<unsigned> DrawCallToMaterialBinding();
+        std::vector<uint64> DrawCallToMaterialBinding();
         void LogReport() const;
 
             ////////////////////////////////////////////////////////////
